Removal of files
Deleted files which should not have been in folder
